components/train_test_split/requirements.txt
panda
scikit-learn